Re-arch and cleanup python pep8 warning
