+++ .github/workflows/flake8.yaml
@@ -1,0 +1,28 @@
+name: Python Flake8 Check
+
+on:
+  pull_request:
+    paths:
+    - 'hub-mirror/**'
+    # Runs at every pull requests submitted in master branch
+    branches: [ master ]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python
+      uses: actions/setup-python@v2
+      with:
+        python-version: '3.x'
+    - name: Install dependencies
+      working-directory: ./hub-mirror
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+    - name: Lint with flake8
+      working-directory: ./hub-mirror
+      run: |
+        pip install flake8
+        flake8 .
