+++ Dockerfile
@@ -4,8 +4,7 @@ RUN apt update && apt install git python3 python3-pip -y && \
   echo "StrictHostKeyChecking no" >> /etc/ssh/ssh_config
 
 ADD *.sh /
-ADD hubmirror.py /
-ADD requirements.txt /
+ADD hub-mirror /hub-mirror
 ADD action.yml /
 
 ENTRYPOINT ["/entrypoint.sh"]
