+++ entrypoint.sh
@@ -10,9 +10,9 @@ mkdir -p /root/.ssh
 echo "${INPUT_DST_KEY}" > /root/.ssh/id_rsa
 chmod 600 /root/.ssh/id_rsa
 
-pip3 install -r /requirements.txt
+pip3 install -r /hub-mirror/requirements.txt
 
-python3 /hubmirror.py --src "${INPUT_SRC}" --dst "${INPUT_DST}" \
+python3 /hub-mirror/hubmirror.py --src "${INPUT_SRC}" --dst "${INPUT_DST}" \
 --dst-token "${INPUT_DST_TOKEN}" \
 --account-type "${INPUT_ACCOUNT_TYPE}" \
 --clone-style "${INPUT_CLONE_STYLE}" \
